Renamed creteTour anddeleted crete inputs component or whatever
