+++ frontend/src/views/CreateTourView.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import React from 'react';
-import CreateTour from '../components/Tours/TourInputs'
+import CreateTour from '../components/Tours/CreateTour'
 
 export class CreateTourView extends React.Component {
